Project summary for Off-school lessons App @naijahacks
PROJECT SUMMARY
Category : Education
Team: Android_geek
App Name: Off-School Lesson App.
Team member: Akinwumi Ayodimeji (Programmer), Bello Habeebllahi Afolabi (Programmer), Sanni Ademola ( Designer).

This project is basically student for student ranging from primary to junior secondary school, As a team, we discuss student between the above range, always leave home very early as 8Am in the morning and arrive home minimum of 4-5pm, and this always outstretched them for the whole day, because of their extra-lesson. As a team we concluded to develop an application that we allow student to have access to lesson, Off-school at their Comfort Zone, questions can be ask by student, and the teacher we give feedback based on the questions; we belief, this we create more time for student to relax their brain and active in other things within the day, also this wevas well, create more time for the teachers, and also add value to any education system that implement this innovation... This Application we be developed using Android Studio(XML, JAVA) with firebase as our database. Thanks to naijahacks for creating this great platform.